+++ includes/Triangle.h
@@ -18,19 +18,19 @@ public:
   ~Triangle2D() { }
 
   const bool contains(const Point2D& point) const {
-    float u, v, w;
+    double u, v, w;
     calculateBarycentricCoords(u, v, w, point);
     return (u >= 0 && v >= 0 && u + v < 1);
   }
 
-  void calculateBarycentricCoords(float& u, float& v, float& w, const Point2D& point) const {
+  void calculateBarycentricCoords(double& u, double& v, double& w, const Point2D& point) const {
     const Vector2D v0 = this->v2 - this->v1, v1 = this->v3 - this->v1, v2 = point - this->v1;
-    float d00 = v0 * v0;
-    float d01 = v0 * v1;
-    float d11 = v1 * v1;
-    float d20 = v2 * v0;
-    float d21 = v2 * v1;
-    float denom = d00 * d11 - d01 * d01;
+    double d00 = v0 * v0;
+    double d01 = v0 * v1;
+    double d11 = v1 * v1;
+    double d20 = v2 * v0;
+    double d21 = v2 * v1;
+    double denom = d00 * d11 - d01 * d01;
     v = (d11 * d20 - d01 * d21) / denom;
     w = (d00 * d21 - d01 * d20) / denom;
     u = 1.0f - v - w;
